Homework #5.1 has been completed.
