+++ package.json
@@ -35,6 +35,7 @@
   "devDependencies": {
     "chai": "^4.2.0",
     "eslint": "^5.8.0",
+    "extract-text-webpack-plugin": "^4.0.0-beta.0",
     "istanbul": "^0.4.5",
     "istanbul-instrumenter-loader": "^3.0.1",
     "karma": "^3.1.1",
