+++ webpack.config.js
@@ -1,5 +1,5 @@
 const HtmlPlugin = require('html-webpack-plugin');
-const CleanWebpackPlugin = require('clean-webpack-plugin');
+// const CleanWebpackPlugin = require('clean-webpack-plugin');
 const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 const rules = require('./webpack.config.rules');
 const fs = require('fs');
@@ -68,7 +68,8 @@ module.exports = {
         new MiniCssExtractPlugin({
             filename: '[name].css',
         }),
-        ...html,
-        new CleanWebpackPlugin(['dist'])
+        ...html
+        // ,
+        // new CleanWebpackPlugin(['dist'])
     ]
 };
